% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{aws_query_string_auth_url}
\alias{aws_query_string_auth_url}
\title{Generating URLs for S3 resources with query string request authentication.}
\usage{
aws_query_string_auth_url(dt = NULL, full_name = NULL,
  region = "us-east-1",
  aws_access_key_id = data.table::fread("~/publicread.access.id.txt", header =
  FALSE)$V1,
  aws_secret_access_key = data.table::fread("~/publicread.secret.key.txt",
  header = FALSE)$V1, lifetime_minutes = 60, lifetime_days = NULL,
  bitly = FALSE)
}
\arguments{
\item{dt}{Data table. Output from index_S3_objects.}

\item{full_name}{Character. Objects to lookup.}

\item{region}{AWS region the bucket is located in.}

\item{aws_access_key_id}{access key ID for an IAM user with access to the S3 bucket.}

\item{aws_secret_access_key}{secret access key for an IAM user with access to the S3 bucket.}

\item{lifetime_minutes}{Numeric. Duration in minutes after which the URL will expire.}

\item{lifetime_days}{Numeric. Duration in days after which the URL will expire.}

\item{bitly}{Logical. Generated a shortened url using bit.ly?}
}
\description{
Generating URLs, optionally shortened, for S3 resources with query string request authentication for the specified time interval.
}
