% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{copy_S3_object_version}
\alias{copy_S3_object_version}
\title{Copying S3 object versions.}
\usage{
copy_S3_object_version(dt, cores = parallel::detectCores() * 4,
  storage.class = "STANDARD", canned.acl = "private", copy.prefix = NULL)
}
\arguments{
\item{dt}{Data table. Output from index_S3_objects.}

\item{cores}{Numeric. Cores to parallelize over.}

\item{storage.class}{Characer. Storage class. One of STANDARD, REDUCED_REDUNDANCY, STANDARD_IA.}

\item{canned.acl}{Character. Canned ACL to apply to the object. One of private, public-read, public-read-write, authenticated-read, aws-exec-read, bucket-owner-read, bucket-owner-full-control.}

\item{copy.prefix}{Character. Prefix to copy to, begining with bucket name. NULL will copy the object in place.}
}
\description{
Copying S3 object versions.
}
