% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{get_S3_object_version}
\alias{get_S3_object_version}
\title{Download S3 object versions.}
\usage{
get_S3_object_version(dt, version = TRUE, cores = parallel::detectCores() *
  4)
}
\arguments{
\item{dt}{Data table. Output from index_S3_objects.}

\item{version}{Logical. Prepend verion to filename?}

\item{cores}{Numeric. Cores to parallelize over.}
}
\description{
Download S3 object versions.
}
