% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{import_S3_json}
\alias{import_S3_json}
\title{Import AWS CLI S3 list-object-versions json output to a data table.}
\usage{
import_S3_json(x)
}
\arguments{
\item{x}{Character. AWS CLI s3api json output.}
}
\description{
Import AWS CLI S3 list-object-versions json output to a data table.
}
