% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{index_S3_objects}
\alias{index_S3_objects}
\title{Index S3 objects from a list of S3 buckets.}
\usage{

  index_S3_objects(buckets = system("aws s3 ls | grep -Eo '[a-z0-9\\\\.]{1,99}$'",
  intern = TRUE), prefix = NULL)
}
\arguments{
\item{buckets}{Character. List of S3 buckets to index.}

\item{prefix}{Character. Prefix filter.}
}
\description{
Quickly index all S3 objects in a bucket list into a data table. Requires AWS CLI.
}
