% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{print_S3_statistics}
\alias{print_S3_statistics}
\title{Print statistics about S3 buckets.}
\usage{
print_S3_statistics(S3_dt, log = "./print_S3_statistics.log")
}
\arguments{
\item{S3_dt}{Data table output from index_S3_objects.}

\item{log}{File path for log output.}
}
\description{
Print statistics about S3 buckets.
}
