% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.s3.R
\name{restore_S3_object_version}
\alias{restore_S3_object_version}
\title{Restore S3 object versions from AWS Glacier.}
\usage{
restore_S3_object_version(dt, cores = parallel::detectCores() * 4, days = 7)
}
\arguments{
\item{dt}{Data table. Output from index_S3_objects.}

\item{cores}{Numeric. Cores to parallelize over.}

\item{days}{Numeric. Days to restore objects for.}
}
\description{
Restore S3 object versions from AWS Glacier.
}
